pyproject.toml
[project]
name = "everbrowser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.5",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-openai>=1.0.2",
    "pillow>=12.0.0",
]

[[tool.uv.index]]
url = "http://mirrors.aliyun.com/pypi/simple"
default = true
